consolidate shared code into _shared
