+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/__init__.py
@@ -3,6 +3,5 @@
 # Licensed under the MIT License.
 # -------------------------------------
 from ._client import KeyClient
-from ._models import Key, KeyBase, DeletedKey, KeyOperationResult
 
 __all__ = ["KeyClient"]
