+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_shared/_generated/key_vault_client.py
