+++ sdk/keyvault/azure-keyvault-keys/azure/keyvault/keys/_shared/_generated/v7_0/version.py
